### 2018-11-01 #1060 @whedon
[PRE REVIEW]: Artery.FE: An implementation of the 1D blood flow equations in FEniCS
**Submitting author:** @akdiem (<a href="http://orcid.org/0000-0003-1719-1942">Alexandra Diem</a>)
**Repository:** <a href="https://github.com/KVSlab/bloodflow" target ="_blank">https://github.com/KVSlab/bloodflow</a>
**Version:** v1.0
