### 2018-11-02 #1060 @trallard
Hi @akdiem I will be the editor for your submission 👋
It seems there are some problems with the paper, you might want to check the `yaml` header and try and regenerate the paper.

Also, have you got any potential reviewers in mind? If not you can check our list [of registered reviewers](https://bit.ly/joss-reviewers)
