### 2018-11-06 ac0a1a1
affiliations were missing from paper.md
