### 2018-11-06 #1060 @akdiem
Hi @trallard Thank you for being the editor for my submission and apologies for the delayed response (I was away). I have fixed the issue according to the error message and pushed the changes to my repository. How do I regenerate the paper?

I have checked the list of reviewers and could suggest @abhishekbajpayee and @dlagrava
