+++ paper/paper.bib
@@ -1,15 +1,19 @@
 @article{Diem:2016,
 	Author = {Alexandra K. Diem, Neil W. Bressloff},
-	Journal = {Journal of Open Research Software}
-	year = {2016}
-	Doi = {10.5334/jors.159}
+	Journal = {Journal of Open Research Software},
+	Year = {2016},
+	Number = {17},
+  Volume = {5},
+	Doi = {10.5334/jors.159},
 	Title = {VaMpy: a Python Package to solve 1D blood flow problems}
 }
 @article{Kolachalama:2007,
 	Author = {Kolachalama, V. B., Bressloff, N. W., Nair, P. B., & Shearman, C. P.},
 	Journal = {IEEE Transactions on Biomedical Engineering},
 	Year = {2007},
-	Month = {May},
+	Number = {5},
+  Pages = {802 -- 812},
+  Volume = {54},
 	Doi = {10.1109/tbme.2006.889188},
 	Title = {Predictive Haemodynamics in a One-Dimensional Human Carotid Artery Bifurcation. Part I: Application to Stent Design}
 }
@@ -17,11 +21,14 @@
 	Author = {Olufsen, M. S., Peskin, C. S., Kim, W. Y., Pedersen, E. M., Nadim, A., & Larsen, J},
 	Journal = {Annals of Biomedical Engineering},
 	Year = {2000},
+	Number = {11},
+  Pages = {1281 -- 1299},
+  Volume = {28},
 	Doi = {10.1114/1.1326031},
 	Title = {Numerical Simulation and Experimental Validation of Blood Flow in Arteries with Structured-Tree Outflow Conditions}
 }
 @article{fenics},
-	Author = {Alnaes, M. S., Blechta, J., Hake, J., Johansson, A., Kehlet, B., Logg, A., Richardson, J., Rognes, M. E., Wells, G. N.}
+	Author = {Alnaes, M. S., Blechta, J., Hake, J., Johansson, A., Kehlet, B., Logg, A., Richardson, J., Rognes, M. E., Wells, G. N.},
 	Journal = {Archive of Numerical Software},
 	Year = {2015},
 	Doi = {10.11588/ans.2015.100.20553},
