+++ paper/paper.bib
@@ -27,7 +27,7 @@
 	Doi = {10.1114/1.1326031},
 	Title = {Numerical Simulation and Experimental Validation of Blood Flow in Arteries with Structured-Tree Outflow Conditions}
 }
-@article{fenics},
+@article{fenics,
 	Author = {Alnaes, M. S., Blechta, J., Hake, J., Johansson, A., Kehlet, B., Logg, A., Richardson, J., Rognes, M. E., Wells, G. N.},
 	Journal = {Archive of Numerical Software},
 	Year = {2015},
