### 2018-11-08 #1060 @trallard
Hi @dlagrava no worries that would be perfect, we ideally want reviews to be conducted within a couple of weeks but taking longer than that should not be an issue, we can always accommodate and work around reviewer's availability. Does this sound reasonable?
