### 2018-11-16 #1060 @trallard
@akdiem  let's ping a few fitting reviewers, with Python and computational fluid dynamics expertise. We already have @dlagrava acceptance to review but ideally, we'd want one extra reviewer.

Dear reviewers,

@kyleniemeyer  Kyle, are you able to and interested in taking on this review?
@abhishekbajpayee Abishek, are you able to and interested in taking on this review?
@ctdegroot Christopher, are you able to and interested in taking on this review?

---

**Title:** Artery.FE: An implementation of the 1D blood flow equations in FEniCS
**Summary:** This package implements the 1D blood flow equations using the finite element framework FEniCS. The package provides tools for modelling blood flow through a network of arteries by solving a 1D approximation of the cross-sectional area and flow rate of each artery.
**Topic:** Computational Fluid Dynamics
**Article Proof:** [https://github.com/openjournals/joss-papers/blob/joss.01060/joss.01060/10.21105.joss.01060.pdf](https://github.com/openjournals/joss-papers/blob/joss.01060/joss.01060/10.21105.joss.01060.pdf)
Submitting author: @akdiem
**Repository:** [https://github.com/KVSlab/bloodflow](https://github.com/KVSlab/bloodflow)
