### 2018-11-17 #1060 @ctdegroot
@trallard I am interested, but I have two other JOSS reviews on the go right now. I am willing to do it, but with the understanding that it might take me a few weeks to start my review.
