### 2018-11-28 #1060 @trallard
Hi @ctdegroot, thanks a lot for offering, but given the fact that you have two simultaneous reviews at the moment I might proceed to review with @dlagrava.

We very much appreciate your time and contribution to JOSS and we would certainly not have you burn out 😄
