### 2018-11-29 #1107 @trallard
Hi reviewers and @akdiem ! This is the issue we will be using to conduct the review. I know @jhill1 is away for a couple of weeks but in the meantime, @dlagrava can get started with the review and recommendations (if any).

If any of you encounters any issues / need any guidance during the review process feel free to ping me here 👩🏻‍💻
