### 2018-11-29 #1107 @dlagrava
@akdiem I have started this. The first comment that I have is that you actually do not have a release v1.0 on your github. Could you please create one? Thanks.
