### 2018-12-03 #1107 @trallard
Thanks for the note @akdiem I will revise the submission guide and make this clearer for future releases
