### 2018-12-03 #1107 @dlagrava
@akdiem I have gone through the installation part. It is really a great idea to provide a Docker container. My only comment at this point is that the `python` command inside the container points to python 2, and you need to specifically need to execute the code with `python3`. Could you please update the readthedocs to reflect this or make the default python to be 3 on the Dockerfile. Thanks.
