### 2018-12-03 #1107 @dlagrava
@akdiem I have tried to run the unit tests that you have provided. The `test_artery_network.py` does not work as the class it reffers to seems to have changed from "Artery_Network" to "ArteryNetwork". Thanks.
