### 2018-12-04 eebd9b9
Changed 'python' to 'python3' in the documentation
