### 2018-12-04 95046e9
corrected reference to ArteryNetwork class
