### 2018-12-04 2731f83
deleted README.rst as it is not used
