### 2018-12-04 a633ac8
fixed reference to ArteryNetwork in tests
