### 2018-12-04 6f4b6f8
minor fixes to command lines
