+++ docs/source/getting_started.rst
@@ -14,11 +14,11 @@ We recommend installing artery.fe using the provided Dockerfile. This
 ensures all dependencies are correctly installed. Build the Docker image
 by running::
 
-  $ docker build --no-cache -t arteryfe:2017.2.0 .``
+  $ docker build --no-cache -t arteryfe:2017.2.0 .
 
 To create and enter a Docker container run::
 
-  $ docker run -ti -p 127.0.0.1:8000:8000 -v (pwd):/home/fenics/shared -w /home/fenics/shared "arteryfe:2017.2.0"``
+  $ docker run -ti -p 127.0.0.1:8000:8000 -v $(pwd):/home/fenics/shared -w /home/fenics/shared "arteryfe:2017.2.0"
 
 Alternatively, artery.fe can be installed using the provided `setup.py` file by running::
 
