### 2018-12-12 KVSlab/bloodflow#38 @jhill1
Error in referencing for JOSS paper
This issue is part of the [JOSS paper review](https://github.com/openjournals/joss-reviews/issues/1107).

There seems to be a formatting issue with a couple of the references:
@Olufssen:2000 and @Kolachalama:2007
which appear in the final sentence of the paper.
