### 2018-12-12 0b56b39
Update paper.md
fixed typo in reference
