### 2018-12-12 KVSlab/bloodflow#39 @jhill1
Add example docs to README
This issue is part of this [JOSS review](https://github.com/openjournals/joss-reviews/issues/1107).

It would be good to have an explicit command to run the example set up, e.g.:
``python3 demo_arterybranch.py config/demo_arterybranch.cfg``
in the README, or at least some help when the user just calls:
``python3 demo_arterybranch.py``
Would help get them started :)
