### 2018-12-12 3dd6a1c
Update paper.bib
fixed author lists
