+++ paper/paper.bib
@@ -8,7 +8,7 @@
 	Title = {VaMpy: a Python Package to solve 1D blood flow problems}
 }
 @article{Kolachalama:2007,
-	Author = {Kolachalama, V. B., Bressloff, N. W., Nair, P. B., & Shearman, C. P.},
+	Author = {Kolachalama V. B., Bressloff N. W., Nair P. B., & Shearman C. P.},
 	Journal = {IEEE Transactions on Biomedical Engineering},
 	Year = {2007},
 	Number = {5},
@@ -18,7 +18,7 @@
 	Title = {Predictive Haemodynamics in a One-Dimensional Human Carotid Artery Bifurcation. Part I: Application to Stent Design}
 }
 @article{Olufsen:2000,
-	Author = {Olufsen, M. S., Peskin, C. S., Kim, W. Y., Pedersen, E. M., Nadim, A., & Larsen, J},
+	Author = {Olufsen M. S., Peskin C. S., Kim W. Y., Pedersen E. M., Nadim A., & Larsen J},
 	Journal = {Annals of Biomedical Engineering},
 	Year = {2000},
 	Number = {11},
@@ -28,7 +28,7 @@
 	Title = {Numerical Simulation and Experimental Validation of Blood Flow in Arteries with Structured-Tree Outflow Conditions}
 }
 @article{fenics,
-	Author = {Alnaes, M. S., Blechta, J., Hake, J., Johansson, A., Kehlet, B., Logg, A., Richardson, J., Rognes, M. E., Wells, G. N.},
+	Author = {Alnaes M. S., Blechta J., Hake J., Johansson A., Kehlet B., Logg A., Richardson J., Rognes M. E., Wells G. N.},
 	Journal = {Archive of Numerical Software},
 	Year = {2015},
 	Doi = {10.11588/ans.2015.100.20553},
