### 2018-12-12 KVSlab/bloodflow#38 @akdiem
I've attemted to fix them, but for some reason that I do not understand yet, the author list on those references is ignored when building the pdf.
