### 2018-12-12 4ee06fc
Update paper.bib
Made author list formatting consistent
