+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Diem:2016,
-	Author = {Alexandra K. Diem, Neil W. Bressloff},
+	Author = {Diem A. K., Bressloff N. W.},
 	Journal = {Journal of Open Research Software},
 	Year = {2016},
 	Number = {17},
