### 2018-12-12 KVSlab/bloodflow#38 @jhill1
I think it's the formatting of your authors, e.g.:

``Alnaes, M. S. and Blechta, J. and Hake, J. and Johansson, A. and Kehlet, B. and Logg, A. and Richardson, J. and Rognes, M. E. and Wells, G. N.``

Note the 'and' and comma placement.
