### 2018-12-12 #1107 @jhill1
Hi @akdiem. I found the package very interesting and nice use of Fenics. The module is very well documented with inline comments, and the documentation has examples, tests. The paper is well structured with a clear aim for the software and the equation laid out are easy to find in the corresponding code. The tests worked out of the box for me and installation was straightforward using the Docker or the more traditional setup.py.

The example given is well described, giving reference to the paper that it uses and the plot produced even use a good colour scheme!

There are a few minor issues with formatting of references in the paper and I would recommend hooking up the tests to TravisCI or similar system for continuous testing. The former should be done before accepting this paper, but the latter is at the discretion of the authors (as there are manually running tests).

Once the issues with the referencing are sorted, this paper should be accepted.

Note to @trallard I'm not an expert in blood flow equations, but the Fenics code looks good to me.
