### 2018-12-13 #1107 @dlagrava
Hello @akdiem I have finished my review as well. I performed some blood simulations for stent optimisation on 3D and it was very time consuming, probably a 1D approach should be more efficient, while retaining most of the average behavior. I quite enjoyed the presentation of the theory of the 1D bloodflow equations on your readthedocs, which is a nice touch for the interested reader.

Your implementation based on FEnICS is well  documented, the installation and the tests work well. One question would be how you manage/inform about updates and bug fixes on FEnICS.

The last comment that I have is that your readme.md on the project should provide guidelines stating that people are welcome to contribute and submit bugs, and who to ask questions to.

It was an interesting review. I wish you a lot of success.

@trallard Once the comment about community guidelines is fixed, I have finished with the review.
