### 2018-12-14 bf6bd4d
Update paper.bib
Removed "&" from author lists.
