### 2018-12-14 e65b56e
Update README.md
Added Usage section
