### 2018-12-14 KVSlab/bloodflow#39 @akdiem
Thank you for the suggestion, I've added a "Usage" section with the example set up to the README.
