### 2018-12-14 f66fb53
Update paper.bib
Authors are split by "and", first and last name by comma
