### 2018-12-14 d03c479
Update paper.bib
added 2 missing commas
