+++ paper/paper.bib
@@ -8,7 +8,7 @@
 	Title = {VaMpy: a Python Package to solve 1D blood flow problems}
 }
 @article{Kolachalama:2007,
-	Author = {Kolachalama, V. B. and Bressloff, N. W. and Nair, P. B. and Shearman C. P.},
+	Author = {Kolachalama, V. B. and Bressloff, N. W. and Nair, P. B. and Shearman, C. P.},
 	Journal = {IEEE Transactions on Biomedical Engineering},
 	Year = {2007},
 	Number = {5},
@@ -18,7 +18,7 @@
 	Title = {Predictive Haemodynamics in a One-Dimensional Human Carotid Artery Bifurcation. Part I: Application to Stent Design}
 }
 @article{Olufsen:2000,
-	Author = {Olufsen, M. S. and Peskin, C. S. and Kim, W. Y. and Pedersen, E. M. and Nadim, A. and Larsen J},
+	Author = {Olufsen, M. S. and Peskin, C. S. and Kim, W. Y. and Pedersen, E. M. and Nadim, A. and Larsen, J},
 	Journal = {Annals of Biomedical Engineering},
 	Year = {2000},
 	Number = {11},
