### 2018-12-14 #1107 @akdiem
@jhill1 Thank you for your review and your comments! I agree that CI use would be ideal. It has been quite tricky so far to set up the specific version of Fenics on Travis as I would need to install it via docker. I will look into this again though.
