### 2018-12-14 1067b66
Update README.md
Added a support section
