### 2018-12-14 #1107 @akdiem
@dlagrava Thank you for your review and your comments! I have added a "Support" section to the README, stating the info you requested. If you would like to explore running your simulations using artery.FE, I'd be more than happy to assist. Feel free to contact me about this.
