### 2018-12-17 #1107 @trallard
Hey no worries @dlagrava thanks again for your time

As soon as @jhill1 confirms that the community guidelines point has been satisfactorily added I will conduct the last editorial tasks on the submission
