### 2018-12-17 #1107 @trallard
@akdiem now that the reviews are completed and the paper has been recommended for acceptance, we need you to complete the following actions:

- [ ]   make a new release of the package (unless the latest release is up to date with the recent modifications)

- [ ]  upload the revised software to your DOI-granting data/software repository, and post the DOI here
