### 2018-12-18 3d0487e
Update paper.md
Added acknowledgements
