+++ paper/paper.md
@@ -49,4 +49,9 @@ where $U = (A, q)$, $F(U) = (q, \frac{q^2}{A} + f(r_0) \sqrt{A_0 A})$, $S(U) = (
 The above system of equations governs the time-development of the cross-sectional artery and blood flow rate in one artery. The associated boundary conditions are as follows: The initial radii respect the relation $r_0(x) = Ru (\frac{Rd}{Ru})^{x/L}$ and the initial cross-sectional area for each artery is thus $A = \pi (r_0(x))^2$. A prescribed flow rate at the inlet of the root artery serves as the inlet boundary condition. For each artery, the initial flow is computed according to the artery's share of the cross-sectional area at the bifurcation. At each bifurcation point, the area and the flow is computed using Richtmyer's two-step Lax-Wendroff method, by solving a system of equations with 18 unknowns. At the terminal arteries, pressure is computed using a Windkessel model and prescribed as the outlet boundary condition. When the pressure is known, the area can be computed from the state equation: $p - p_0 = f(r_0) (1 - \sqrt{\frac{A_0}{A}})$. The derivation of the governing system of equations and boundary conditions can be found in [@Olufsen:2000, @Kolachalama:2007, @Diem:2016].
 
 
+# Acknowledgements
+
+The authors acknowledge the European Community for its financial support for KVS and AKD in the framework of the project CUPIDO (www.cupidoproject.eu) H2020-NMBP-2016 720834, as well as the Simula Research Laboratory for financial support for SDA.
+
+
 # References
