### 2018-12-18 #1107 @akdiem
@trallard I had forgotten to add the acknowledgements to the paper - I hope it is ok to still add them. I have rebuilt the paper and will perform the other final steps now.
