### 2018-12-18 ef2ac06
Update README.md
corrected doi to publication release
