+++ README.md
@@ -1,4 +1,4 @@
-[![DOI](https://zenodo.org/badge/138987449.svg)](https://zenodo.org/badge/latestdoi/138987449)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2383815.svg)](https://doi.org/10.5281/zenodo.2383815)
 
 # artery.fe
 
