### 2018-12-18 #1107 @trallard
@arfon: this submission is accepted and ready to be published 🎉👾

@jhill1  and @dlagrava  thank you very much for your time and valuable contribution to JOSS as a reviewer for this submission 🙌🏻🌟
